requirements.txt
flask
flask-cors
torch
torchvision
pillow
opencv-python
mediapipe
numpy
pandas
requests
matplotlib
